namespace folders into 'metadata'
