--- offchain/adapters/http_adapter.py
@@ -1,7 +1,0 @@
-from offchain.adapters.base_adapter import HTTPAdapter as BaseHTTPAdapter
-from offchain.registries.adapter_registry import AdapterRegistry
-
-
-@AdapterRegistry.register
-class HTTPAdapter(BaseHTTPAdapter):
-    pass
