+++ offchain/metadata/adapters/base_adapter.py
