+++ offchain/metadata/models/token.py
@@ -1,0 +1,8 @@
+from offchain.base.base_model import BaseModel
+
+
+class Token(BaseModel):
+    chain_identifier: str
+    collection_address: str
+    token_id: int
+    uri: str
