+++ offchain/metadata/registries/base_registry.py
